+++ src/ui/central_panel.rs
@@ -187,6 +187,10 @@ impl WalksnailOsdTool {
                 ui.label("Encoding bitrate");
                 ui.add(Slider::new(&mut self.render_settings.bitrate_mbps, 0..=100).text("Mbps"));
                 ui.end_row();
+
+                ui.label("Upscale to 1440p");
+                ui.add(Checkbox::without_text(&mut self.render_settings.upscale));
+                ui.end_row();
             });
     }
 }